chore(overview): Move context logic to hook file
